+++ changelog/mediation/mobileads-digitalturbine/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Version 8.3.2.4
+
+### Updated
+
+- Updated minimum supported `YandexMobileAds` version to `7.14.1`
+
 ## Version 8.3.2.3
 
 ### Updated